move schema checker
